+++ source/Layers/LayersPanel.ts
@@ -1,14 +1,19 @@
 class LayersPanel implements UIElement {
-    private container: HTMLDivElement;
+    private readonly container: HTMLDivElement;
+    private readonly layers: Dictionary<HTMLElement> = {};
 
-    public constructor(private layers: LayersManager, private localizer: Localizer) {
+    public constructor(
+        private layersManager: LayersManager,
+        private uiFactory: UIFactory,
+        private localizer: Localizer
+    ) {
         const container = document.createElement('div');
 
         container.className = 'layers';
 
-        this.layers.onCreate(l => this.buildLayer(l));
-        this.layers.onDelete(l => this.removeLayer(l.id));
-        this.layers.onSelect(l => this.selectLayer(l));
+        layersManager.onCreate(l => this.buildLayer(l));
+        layersManager.onDelete(l => this.removeLayer(l.id));
+        layersManager.onSelect(l => this.selectLayer(l));
 
         this.container = container;
     }
@@ -20,14 +25,15 @@ class LayersPanel implements UIElement {
     private buildLayer(layer: LayerAccessor) {
         const data = layer.data,
             id = data.id,
-            mapId = this.layers.mapId,
+            mapId = this.layersManager.mapId,
             //HACK: Magic string layer_type_
             type = this.localizer[`layer_type_${data.type}`],
             wrapper = document.createElement('div'),
             check = document.createElement('input'),
             radio = document.createElement('input'),
             label = document.createElement('label'),
-            typeLabel = document.createElement('small');
+            typeLabel = document.createElement('small'),
+            deleteButton = this.uiFactory.createCloseButton(_ => this.layersManager.delete(id));
 
         check.type = 'checkbox';
         check.name = mapId + '-visible-layers';
@@ -35,7 +41,7 @@ class LayersPanel implements UIElement {
         check.id = id + '-visible';
         check.checked = !data.hidden;
 
-        check.onchange = () => this.layers.update(id, l => l.hidden = !check.checked);
+        check.onchange = () => this.layersManager.update(id, l => l.hidden = !check.checked);
 
         radio.type = 'radio';
         radio.name = mapId + '-active-layer';
@@ -43,7 +49,7 @@ class LayersPanel implements UIElement {
         radio.id = this.getRadioId(id);
         radio.className = 'label-radio';
 
-        radio.onchange = () => this.layers.select(id);
+        radio.onchange = () => this.layersManager.select(id);
 
         typeLabel.innerText = `(${type})`;
 
@@ -51,26 +57,22 @@ class LayersPanel implements UIElement {
         label.innerText = data.name;
         label.title = data.name;
         label.append(typeLabel);
-
-        wrapper.id = this.getWrapperId(id);
+        label.append(deleteButton);
 
         wrapper.append(check);
         wrapper.append(radio);
         wrapper.append(label);
         this.container.append(wrapper);
+        this.layers[id] = wrapper;
     }
 
     private getRadioId(id: string) {
         return id + '-active';
     }
 
-    private getWrapperId(id: string) {
-        return id + '-panel-controls';
-    }
-
     private removeLayer(id: string) {
-        id = this.getWrapperId(id);
-        document.getElementById(id)?.remove();
+        this.layers[id].remove();
+        delete this.layers[id];
     }
 
     private selectLayer(layer: LayerAccessor) {