+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 
+## [alpha.3] - UNRELEASED
+
+### Changed
+
+- UI:
+  - **Layers Panel**:
+    - Layers have a "close" button
+
+
 ## [alpha.2] - 2026-03-29
 
 ### Added